admin_layer/app/requirements.txt
flask
opencv-python
psycopg2
imutils
sklearn
pika
Pillow
rawpy
imageio
pandas
ExifRead
ffmpeg-python